Poetry-in-Nix structure
This lays the groundwork for a Poetry application packaged with Nix.
Using Nix flakes and Poetry we can be sure, with lock files, that
contributers all share the same outputs.

It includes a flake template file for a Poetry application with the
addition of a "devShell" that includes Python 3.9.9 and Poetry 1.1.12.
If a contributor has the latest Nix unstable, they can run `nix develop
--extra-experimental-features flakes --extra-experimental-features
nix-command` to run the exact same Python and Poetry versions included,
even if their system Poetry and Python differs.

The project was previously initialized with `poetry init`, which does
not create the tests folder and basic folder structure of a best
practices Python environment. This commit adds the missing structure
from `poetry new`.