Return JSON object to HTTP endpoint
This emits a JSON object through the Flask HTTP endpoint. Most of the
debugging around this was due to an import error that only arose when
starting a Flask webserver with `flask run`: `Error: While importing
'little_bits_of_buddha.wsgi', an ImportError was raised`. I determined
that `__init__.py` was the problem and moved the logic out to a new
`data.py` file and deleted the `__init__.py` file as a temporary stopgap