Test random_sutta
This test assumes pytest as our testing framework. With Jack's guidance,
I have changed `random_sutta` to take an argument, which when called
with no arguments uses our default dictionary, `data` but can otherwise
be called with our own selected portion of the whole, which allows us to
test using the `assert` keyword.